Finish filtering and aggregating to a sane looking dataset
Use only the questions we need, restructure the data to put all the questions into a single observation for each blockgroup, increase the sample size by changing the threshold for non-walkable to 40th percentile and walkable to 60th percentile. Rename variables and document the final variable set that will be use.